+++ src/com/ft/jenkins/git/GitTagInfo.groovy
@@ -1,4 +1,4 @@
-package com.ft.up.git
+package com.ft.jenkins.git
 
 public class GitTagInfo implements Serializable {
   String authorEmail
