+++ src/com/ft/jenkins/git/GithubReleaseInfo.groovy
@@ -1,4 +1,4 @@
-package com.ft.up.git
+package com.ft.jenkins.git
 
 class GithubReleaseInfo implements Serializable {
   String title
